[pagai] Fix unit tests
